pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "reservoirinvest", email = "reservoirinvest@gmail.com"},
]
dependencies = [
    "jupyter>=1.0.0",
    "lab>=8.0",
    "pandas>=2.1.1",
    "lxml>=4.9.3",
    "pyyaml>=6.0.1",
    "loguru>=0.7.2",
    "ib-insync>=0.9.86",
    "from-root>=1.3.0",
    "tqdm>=4.66.1",
    "yahooquery>=2.3.7",
    "nsepython>=2.8",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

